Create a new component that displays an ajax loader image i.e. used on chapter add

+++ src/main/java/com/wooki/domain/biz/ChapterManagerImpl.java
@@ -237,11 +237,12 @@ public class ChapterManagerImpl extends AbstractManager implements ChapterManage
         List<Publication> publications = publicationDao.listPublication(chapterId);
         if (publications.size() <= 1) { throw new IllegalStateException(
                 "Chapter must be associated at least to one revision"); }
-        
+
         Publication toDelete = getRevision(chapterId, revision);
+        if (toDelete.isPublished()) { throw new IllegalStateException(
+                "Current published revision cannot be deleted"); }
+
         publicationDao.delete(toDelete);
-        
-        Publication last = publicationDao.findLastRevision(chapterId);
     }
 
     public Chapter update(Chapter chapter)
